requirements.txt
streamlit
pandas
pyodbc
bcrypt
python-dotenv
matplotlib
python-pptx
